HistogramBuilder moved to utils
